# OpenAPI.yaml
openapi: 3.0.0
info:
  title: MY SWEET SKRIPSI
  contact: {}
  version: "1.0.0"
servers:
  - url: http://localhost:8000/v1/api
  - url: https://be-permission.up.railway.app/v1/api
    variables: {}
paths:
...
  /room/byid/{id}:
    get:
      tags:
        - For Student Room's
      summary: GET DATA ROOM BY ID
      description: ""
      operationId: GETDATAROOMBYID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful Get Rooms
          headers: {}
          content: {}
        "400":
          description: Error or not Success get data
          headers: {}
          content: {}
        "500":
          description: internal server error
          headers: {}
          content: {}
      deprecated: false
      security:
        - bearerAuth: []
...
  /permission/all:
    get:
      tags:
        - For Data Permission
      summary: GET DATA PERMISSION ALL
      description: ""
      operationId: GETDATAPERMISSIONALL
      parameters: []
      responses:
        "200":
          description: Successful get data permission
          headers: {}
          content: {}
        "400":
          description: Error or not success get data permission
          headers: {}
          content: {}
        "500":
          description: internal server error
          headers: {}
          content: {}
      deprecated: false
      security:
        - bearerAuth: []
  /permission/byid/{id}:
    get:
      tags:
        - For Data Permission
      summary: GET DATA PERMISSION BY ID
      description: ""
      operationId: GETDATAPERMISSIONBYID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful get data permission by id
          headers: {}
          content: {}
        "400":
          description: Error or not success get data permission by id
          headers: {}
          content: {}
        "500":
          description: internal server error
          headers: {}
          content: {}
      deprecated: false
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security: []
...